Refactor code with IntArray and StringArray
